+++ packages/stream_chat_persistence/test/src/mapper/pinned_message_mapper_test.dart
@@ -131,7 +131,7 @@ void main() {
     expect(message.user!.id, entity.userId);
     expect(message.localDeletedAt, isSameDateAs(entity.localDeletedAt));
     expect(message.remoteDeletedAt, isSameDateAs(entity.remoteDeletedAt));
-    expect(message.deletedOnlyForMe, entity.deletedForMe);
+    expect(message.deletedForMe, entity.deletedForMe);
     expect(message.text, entity.messageText);
     expect(message.pinned, entity.pinned);
     expect(message.pinExpires, isSameDateAs(entity.pinExpires));
@@ -220,7 +220,7 @@ void main() {
       user: user,
       localDeletedAt: DateTime.now(),
       deletedAt: DateTime.now().add(const Duration(seconds: 1)),
-      deletedOnlyForMe: true,
+      deletedForMe: true,
       text: 'Hello',
       pinned: true,
       pinExpires: DateTime.now(),
@@ -259,7 +259,7 @@ void main() {
     expect(entity.userId, message.user!.id);
     expect(entity.localDeletedAt, isSameDateAs(message.localDeletedAt));
     expect(entity.remoteDeletedAt, isSameDateAs(message.remoteDeletedAt));
-    expect(entity.deletedForMe, message.deletedOnlyForMe);
+    expect(entity.deletedForMe, message.deletedForMe);
     expect(entity.messageText, message.text);
     expect(entity.pinned, message.pinned);
     expect(entity.pinExpires, isSameDateAs(message.pinExpires));