+++ packages/stream_chat_persistence/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Upcoming Beta
 
-- Added support for `Messages.deleteForMe` and `Members.deletedMessages` fields.
+- Added support for `Messages.deletedForMe` and `Members.deletedMessages` fields.
 
 ## 10.0.0-beta.6